examples/example_1/requirements.txt
sqlalchemy
cardda_python
python-dotenv
